weight in gramms

+++ open/src/main/java/com/lenta/bp18/features/good_info/GoodInfoFragment.kt
@@ -34,7 +34,7 @@ class GoodInfoFragment : CoreFragment<FragmentGoodInfoBinding, GoodInfoViewModel
         provideViewModel(GoodInfoViewModel::class.java).let {
             getAppComponent()?.inject(it)
             it.selectedEan.value = goodParams.ean
-            it.weight.value = goodParams.weight.toInt()
+            it.weight.value = goodParams.weight.toDouble()
             return it
         }
     }
